Use hashes in deps dir name
This way we keep track of the version of a dependency since the hash is
the ultimate identity of a version. Previously, if we updated a
dependency so its hash would be updated, we couldn't detect that since
the local dependency copied into .build/deps was just identified by name.