Replace C NULL for iter stop with StopIteration
Instead of returning a C NULL when we want to stop iteration, we now
raise the StopIteration exception. This is a prerequisite for doing
unboxed values since we need the NULL value for unboxed values.

Longer term, we might go back to using NULL but packed in a tuple, like
a tagged union basically, which is likely faster than raising an
exception (that uses longjmp) but that requires some other support.