Stop dependency compiler sub-processes on error
If we encounter errors in building dependencies, we now exit the main
process, but not before stopping our child processes, i.e. the compiler
sub-processes we use to compile our dependencies. We don't wait for the
stopping to complete though, just blindly send some signals. This could
potentially be improved further.