Make it possible to override dep path
The build.act.json of a project specifies the URL and hash of
dependencies. Sometimes it is useful, in particular when doing
development of a project simultaneously as improving one of its
dependencies, that we want to use a local copy of a dependency. Changing
build.act.json and add a path is one way, but it's a little bit awkward
and I have ended up mistakenly committing the local "path": attribute
more than once.. so here is an alternative, we can now specify
depeednncy path overrides to acton using the --dep argument, e.g.

  acton build --dep foo=../foo --dep bar=/some/path/to/bar

Which means that the dependency named foow ill used the local path
../foo instead of whatever is specified in build.act.json. The path can
be absolute or relative. The dep paths are passed down, so if one of our
deps has in its turn a dependency on 'foo' then it too will use the path
override. We pass around absolute paths so that it works regardless of
where the dependencies are.