Reverse order of tests
We construct the lists by prepending using :, which is more efficient
O(1) in Haskell than appending O(n), but it ends up reversing the list
of the tests and I think it's nice if we present the lists to the user
in the same order in which they are defined in the source file. It is
still grouped by test type, but at least there is some more semblence
between list of tests and source file now.