Add __zero__ & use for sum() start arg
The start argument for sum() defaults to None / NULL, which was really
always broken, since not providing led to segfault. We now introduce a
new __zero__ method that is part of the Plus protocol, that is meant to
return the zero value, like an actual 0 for int and other integers or ""
for str. We use it so we don't need to provide an explicit start
argument to sum()