Reorganize compiler into multiple packages
We got a main library and then actonc is a package with just the
executable. This allows us to maintain the versioning we are currently
using, where the actonc executable gets a unique build version for tip
builds. Not sure how important this actually is to maintain, but
multiple packages didn't seem like a big thing either.