Example LSP
The placeholder LSP server has now been replaced with an example LSP. It
still does nothing for Acton but it now does LSP-things, like being able
to start. Using the very latest version of LSP lib since that is what
the example is written for. I have the feeling this is a fast-moving
area, so might as well be on latest and greatest.