Run individual tests, not whole modules
This changes the test running strategy. The top process used to invoke a
whole module, asking it to run all of its tests. In fact, it would run
multiple modules concurrently and within each module, individual tests
would also be run concurrently. It's like two layers of schedulers.

We now move to only invoke a single test from the top process. First we
ask all modules to list all their tests so we get a complete list, after
we start going through test by test, running each as a subprocess. The
module local test runner remains the same, and we select which test to
use by passing the --name argument (which was already available since
before).

We get better control over concurrency, since we can now take decisions
from a single central place. Testing granularity improves, now if a test
crashes with a seg fault or similar, we can report that as an error for
the individual test and not for the whole module as before (which also
required specialized code).

Performance testing should be better and more deterministic since the GC
is not influenced from other tests, since there is only one running at a
time.

A theoretical downside is the cost of spawning more subproceses, but
that should be negligible in comparison to everything else so I won't
even consider it a downside.