Add test timeout
There is now a timeout. It's the test runner that starts a timer. While
it would be nice and sort of self-contained if we could place this in
the test executor itself, that won't work since it might actually be
busy running code. For async tests, it would be fine, but for sync actor
or unit tests, the test function itself is being run and so even if we
schedule a after X callback, it cannot preempt the currently running
test function.

So it's in the test_runner and once it triggers, it emits a message to
the parent process which will register it as a failure. At this point we
also stop the child process. This means we free up resources and can run
the next test.