Simplify test_runner by only running single test
Instead of having the ability to run a list of tests, we can now just
run a single test in the test_runner, which in turn spawns a single
TestExecutor. Since we handle all the concurrency in the parent acton
test process, there is no reason to keep duplicate functionality.

At some point I'd like to bring back concurrency here, but with a
different twist; I'd like to run the same test function by two separate
TestExecutors, so we can check for concurrency issues. This really only
happens in C code, but it's a good test to have.