Capture std(out|err) from test & display
We now capture stdout / stderr from when we run our tests. The recent
change to run each test in its own subprocess effectively makes this
possible since previously the stdout / stderr output from test would be
interleaved from different tests. Now we can neatly capture it all
separately per test.

However, since we run multiple test iterations we now get the output of
multiple test iterations. Need to revise this. We should only show the
first / last or something like that. This is actually sort of hard I
think because stdout output tends to be buffered while stderr is not. We
currently print our test-specific JSON output to stderr, e.g. it is
in-band and get comingled with normal program output over stderr,
however, we print it in such a way that each JSON message should go to
its own line and thus we can separate it out on the receive side.
Printing a magic marker on stderr wouldn't work to determine stdout
output, since that is buffered and printing magic markers on stdout
intermingles even more with normal program output. Separate magic
markers on stdout / stderr could work but is harder timing-wise, since
the buffered stdout would lag behind stderr where we receive JSON
messages.

We should move our test-signaling out-of-band, like send it over a pipe,
but we couldn't do the same for magic markers for test-start / test-end.
I think the only proper way of doing this is to run the test subprocess
with 1 iteration and capture that. That requires a bit of rework in the
top test runner though.