Discover root actor based solely on name
We used to discover the root actor on having both the name "main" as
well as a particular signature, i.e. it should take Env as the single
argument.

The discovery is now only based on the name "main". The root actor is
naturally still expected to take Env as argument, but it now becomes a
type error instead if there is a misalignment. This is more explicit as
otherwise the effect is just that we don't output a binary at all.