Always compile with zig build & remove builder
The Zig build system consists of build.zig, which is a Zig source file.
Before it can run the build, it needs to compile itself. This takes a
few seconds and is normally cached but for temporary acton projects, the
cache won't work. Our solution was to prebuild a 'builder' binary, but
since Zig uses loads of absolute paths for build.zig, we had to add a
couple of hacks to make our builder portable across locations. This
introduced subtle (and sometimes not so subtle) bugs related to caching,
causing invalid use of the cache. Thus, we now remove our builder to
avoid these horrid cache problems.

build.zig files are now always written to disk and kept, so the
--keepbuild option is redundant and thus removed.