Use scratch dirs for single file compilation
Now that our builder has been removed, we suffer from long compilation
times of single .act files since using a temporary directory means the
builder application itself needs to be compiled which takes multiple
seconds.

Instead of creating a temporary directory, we now use a scratch
directory in ~/.cache/acton/scratch. Thus the build.zig binary can be
cached and we get fast single file builds. In order to achieve some
concurrency, we use multiple scratch directories. We attempt to take a
lockfile starting from 0..32 and grab the first one available, so let's
say there are currently 3 concurrent compilations running already, then
we'd get ~/.cache/acton/scratch/scratch3.lock and would then use
~/.cache/acton/scratch/scratch3 as our temporary project directory. We
clean out the scratch project directory after we're done but zig is
smart enough to realize the build.zig remains the same and is cached. On
a M3 MacBook Air, it takes about 0.2 seconds to compile helloworld and
similar small programs. It takes less than a minute to run the whole
test suite when it is cached.

We (ab)use the --tempdir option for this, which we might consider
removing as an exposed option. It was primarily a way to let acton write
the intermediate C output to a known directory, which was a challenge
with random temporary directories. Not sure if it is still needed.