Install headers to avoid zig cache issues
We've been suffering from zig cache issues for quite some time now and
it seems I've finally gotten closer to what might be the culprit. Or at
least a workaround. Installing header files individually instead of
adding a whole directory full of header files, appear to fix our issue.

I've only attempted this with a rather simple reproduction case. It does
fix it entirely, like 100% reproducibility, but I'm not sure that it is
representative of other tests, so I won't claim coverage. Plus, zig
really shouldn't behave like this. Can we be the first to run in to this?

Anyway, we can iterate and come back to this. This seems to improve
things so let's move forward and if it seems to make a true difference
we can dig a little deeper.