Let iN overflow
Our signed integer types are now allowed to overflow. We use the straight
C types and in the C specification, signed integer overflow is undefined
behavior. Most C compilers are quite lax and don't enforce it in any
way per default. We use Zig to compile and per default it enables all
sorts of UBsan, and so we have inherited that behavior.

However, in Acton we think the sane behavior is for integer operations
to be well defined and not change with compilation flags, plus we want
decent performance and so overflow is the most sane behavior. That is
now the case!

Some simple test cases are added to verify overflow works. u16 / i16
does not work, likely since the underpinning type is actually not 16 bits.