Allow dep path OR url in build config
We used to allow path and url, in which case the path would override the
url and this was useful to override with like a local fork of a
dependency when doing development across multiple projects
simultaneously. However, it's bad changing checked in files like
build.act.json and so we added acton build --dep to override deps. It
also has the benefit of being propagated to the build of dependencies,
so if we have a main project that depends on foo and bar. In turn, foo
depends on bar, so if we now override the path to bar, the same override
is passed to the build of foo as to the main project.

Given that we have --dep, the old way of manually overriding the path in
the configuration file is deprecated. We now block and require that only
path OR url is set. Otherwise the user gets an error. We hint about the
--dep argument and this is also based on the config the user has in
their configuration file, e.g.:

  kll@Kristians-MacBook-Air test_deps % cat build.act.json
  {
      "dependencies": {
          "a_b": {
              "path": "deps/a-b",
              "url": "https://github.com/asdf/fnafna/refs/heads/main.zip"
          }
      }
  }
  kll@Kristians-MacBook-Air test_deps % acton build
  ERROR: Dependency 'a_b' has both path and url set. Set only path or url. You can override path for a dep, e.g.: acton build --dep a_b=deps/a-b
  kll@Kristians-MacBook-Air test_deps %