Fix acton test --name
We only expect results from modules that contain a test with the
provided name. Before this fix, acton test --name foo hangs if there are
multiple modules where the name filter only matched on a test in one
module. We would still expect output from the other modules, which would
never come and thus we'd hang.