Add 'acton pkg upgrade'
We now support simple upgrades of pkg dependencies. Running 'acton pkg
upgrade' will upgrade the pkg dependencies configured in build.act.json.
It only supports updating a git branch reference to the HEAD of that
branch. E.g. with a config like this, note the repo_url attribute:

    {
        "dependencies": {
            "foo": {
                "repo_url": "https://github.com/actonlang/foo.git",
                "url": "https://github.com/actonlang/foo/archive/275323b45c318ef600c1b016e17c4337e4186a40.zip",
                "hash": "1220cd47344f8a1e7fe86741c7b0257a63567b4c17ad583bddf690eedd67a032abdd"
            }
        },
        "zig_dependencies": {}
    }

We can upgrade by resolving what the branch HEAD currently points to,
which updates the URL and we also download + recompute the hash and
update build.act.json based on that. Since there is no branch specified,
it uses the default branch (main on the foo repo in this example).

It is possible to specify a different branch using a URL fragment, like so:

    https://github.com/actonlang/foo.git#foobar

This would upgrade to HEAD of the foobar branch.

This current work is based on the intent of the user to use a branch
like main for a dependency. This is fundamentally incompatible with the
Acton view of package dependencies where the hash is really the primary
way of specifying what source we want. The hash is the hash of the
content in the repository, so it is a proxy for saying "we want exactly
this content". It is conceptually more specific than git commit. And we
certainly can't use a branch since the HEAD of the branch changes, and
thus the branch, as a reference mutates and then there's a new hash
which produces a hash mismatch. However, we can't download a hash. We
need some other reference, and that's where git commits come in. Now,
it's pretty painful to manually track git commits and manually add +
recompute the hash for upgrades, so that is what this new 'acton pkg
upgrade' actually solves. In the future we will add support for git tags
as well, so the user can be prompted to upgrade to a particular tag.