Write complete build.zig.zon from cli
actonc previously rewrote parts of build.zig.zon, which is confusing, so
like the acton CLI would write build.zig.zon but then expect actonc to
replace the string "SYSPATH" for the actual path - weird spaghetti
control flow!!

Now, acton directly writes build.zig.zon correctly, so it's usable
straight away. The behavior of actonc is now more explicit and dependent
upon if it is invoked as a sub-compiler (using --sub), i.e. the acton
CLI frontend is the top compiler and invokes actonc in under, in which
case actonc will not meddle with build.zig(.zon) at all. If on the other
hand, actonc is invoked as a standalone compiler, it will write
build.zig(.zon). All code paths means we have up-to-date build.zig(.zon)
- no risk of stale files any longer.