free() is noop!
We should always do a noop_free. I think GC_free() is for leak
detection, but we never use that so for all normal code we should just
never free and instead let the GC do its job.

This was prompted by free bugs related to TLS. I have not verified that
this particular change fixes it, but I believe it is the correct thing
to do.