Guide: side-by-side example + better integer docs
This creates a new side-by-side view for code where the explanatory text
goes alongside the code itself, which I think is a real improvement over
comments or splitting up the code into chunks!

Also improved integer docs a bit and fixed a linking issue where we had
two headers that linked to the same primitivies.md page which led to a
link loop.