Make module path for type error location "clickable"
Use the real filesystem path to the module where a type error was
raised.  This means replacing "." with path separator and appending the
".act" suffix. So foo.bar@123 becomes foo/bar.act@123

This makes it possible for VS Code to create a clickable link to the
file. Well, almost, because VS Code does not recognize "@" as a file -
location separator. Not sure if the diagnostic package can be tweaked to
use ":" instead of "@"?