Silence build lock for dependency builds
It is quite normal to have like ~diamond shaped dependencies in project,
so the top project depends on A & B, and A also depends on B, so when we
start sub-compilers to build dependencies, they will race to build B.
One will make it first and the other will have to wait. It is
unnecessary to print that build lock warning, it only confuses
users (who think something has hung and aborts the whole build).
Detecting this via TTY might not be ideal but it works for now.