Disable GC for acton pkg upgrade
The TLS lib doesn't interoperate well with our GC. While tlsuv does
allow custom mallocs, there is an asymmetry for where malloc / free is
done for DNS addrinfo structs. We control the malloc side and point to
GC_malloc while on the free side, a normal free() is used. That leads to
crashes, so if the GC runs at the wrong time, we'll crash.

Since pkg upgrade is a very small and simple operation, memory usage is
negligible and thus we can live without GC. It's ugly, but at least pkg
upgrade works.

Clearly the solution is to fix the DNS resolution. We already have this
working for TCPConnection, so should port that to TLSConnection.