Await EOF for RunProcess stdout/stderr
Output from the process we run is buffered and can arrive after the
process has exited, so ensure that we await EOF on stdout and stderr
before we actually call it quits.

The callback now takes a ?bytes instead of bytes, so when EOF is reached
we get a None back.

test_process_many has been updated to check the output of the
subprocess. On my mac I can't reproduce the problem so can't tell if
this fixes it, but I think it should.