Ensure worker thread stack size is at least 8MB.
On Mac OS the default stack size for secondary threads is 512 KB, which is too
low for the recursion depth required by the acton-yang generators. On Linux the
default is 1 MB. To ensure we don't get exhaust the stack with recursion again,
we now require the workers to have at least 8 MB -
max(primary_thread_stack_size, 8 MB)

If the primary thread stack size is less than required, then log a warning.