Handle error during output deduplication
I'm not sure when this happens but I have gotten an exception thrown for
a IndexError and this is the only place I can find. After this fix I can
no longer reproduce (though reproduction case was probably <1% in the
first place, so a little hard to tell). I suppose we don't have enough
test output yet, but I don't really get why because we should print \n,
so I think there should always be a second part that is at least the
empty string... but I guess not. Doesn't hurt to be defensive.