Add AsyncTest
This is meant to replace the existing AsyncActorTest. The main idea is
to inject a test-arg object, of type AsyncT, instead of multiple loose
arguments. We can encourage users to explicitly type the arg type and
thus the test discovery becomes more robust. The AsyncT object has
methods for signaling the result:
 - .success(output: ?str=None) for a successful test, possibly with
   golden output
 - .failure(exception: Exception) for a failing test
 - .error(exception: Exception) for an erring test

Once we're happy with this, we will rewrite the other test types to
follow the same style and remove the current ones.