Add new test suite for compiler + f-string tests
This adds a new test suite for the compiler. Previous testing has pretty
much revolved around the idea of first compiling the compiler into an
executable (actonc) and then running that binary executable on a .act
source file to produce some output. This on the other hand is an actual
unit test where we import the parser and call it directly, with all the
benefits that bring.

All the test cases are focused on the new f-string implementation. We
can expand with other tests as we go along.