Parse as stmt, not expr
expr allows too much, so in this case for the error test cases we
actually get parsed output back and not an error as we'd really expect.
Changing to parse stmt gives us an error. We could consider P.file_input
but then we get a tuple back, so I think this is more straight forward
for the vast majority of cases.

This gives us actual error output for out golden tests. The error
message is still rather non-sensical but that has to do with our
f-string parsing rather than anything else.