Add tests for all compiler passes
This acts as a more complete example for how we can write compiler test.
I unified so that the source code for all compiler unit tests are in the
same directory, which I think makes it easier to reuse same source code
for tests that target different passes as well as avoiding that the name
"deact" could mean different tests (source code) for different compiler
passes.