Rename to hash
Follow the style of Rust Hash trait, Swift protocol etc. We'll
eventually stop having a __hash__ function and rely on this
bytes-putting style but since there is just this single way of dealing
with it there is no point using a longer name like hash_bytes when hash
is short and perfectly fine!