Upgrade to Zig v0.14
Pull in new versions of deps modified to work with Zig v0.14

Zig v0.14 has a new fingerprint field in build.zig.zon that is
mandatory. We don't really care about this but we have to write
something so we currently generate a random one - seem to work and I
don't think it impacts caching etc.

Not entirely sure why, but it seems we need to use apple_m1 CPU on MacOS
to get MbedTLS to compile now.