Unescape docstrings in NameInfo
When we create NameInfo entries in Types we bring over docstrings from
the AST. Since we store strings in the AST in an escaped form, we
unescape them when storing as docstrings. In the future, if we decide to
parse docstrings already in the parser AST, we could unescape at that
point and would no longer need extract / unescape here in Types.hs

Also don't add extra whitespace in prettyDocstring when printing this.