Add documentation generation system
Add a new `actonc doc` command that displays documentation for a module
or project. Project documentation is built as part of the normal project
build. Individual modules can be generated on demand. The documentation
is built by extracting docstrings and type information.

- HTML, ASCII (colored) & Markdown formats
- HTML format includes:
  - Clickable type links to definition (cross-module)
  - Tooltips for generic types explaining constraints
  - Dark / light theme, automatic based on OS / browser preference
  - Module index for projects

If we detect a graphical environment, `actonc doc` when run in a
project, will open the project module index in a browser. Otherwise we
display terminal output, which can also be forced using `-t`.

- New DocPrinter.hs module for printing
  - This is largely written by AI. It is quite large (>2K LOC). Focus
    right now are on features. Using AI we can quickly prototype. Once
    we are happy with how docs look, we will clean up the DocPrinter
    module. Since DocPrinter is peripheral in the system, this is
    acceptable.
  - For example, we're experimenting with tooltips on generic types,
    cross-module links to types and more. This will receive more polish
    and we'll converge on a polished design.
- Test modules with golden testing & test/test_doc_printing example
- Uses open on Macos and xdg-open on Linux to open docs in browser