Implement docstring parsing
Docstrings are basically just lone string statements, like a string but
not assigned to anything. It just happens that if a lone string come
just after a declaration, they mean "docstring" and not "string".

We previously extracted docstrings from the AST in the Types pass and
brought these over into a field in NameInfo, in TEnv (type env). It's
now moved to the Parser so that we immediately move over docstrings and
store them separately in the AST. This is done by letting suite return a
tuple of the suite body and the docstring. We save the docstring in a
new ddoc field for all Decl, i.e. functions, classes, actors, extensions
and protocols. In other places we ignore it.

Strings are stored internally in an escaped format. Since docstrings are
docstrings, we can unescape them directly.

In Types, we now lift over the docstring from the AST to NameInfo / TEnv
directly.