Support test discovery of actors
General revamp of test discovery, with the most notable new feature
being discovery of test actors, so we don't have to write stupid wrapper
functions. Now the compiler can write wrappers for us instead!

The test runner takes lists of tests to run. Since the lists need to
have a type but actors do not support inheritance, we cannot have a
common super type. Thus, when we discover a test actor, we create a
wrapper function and use that.

The old test types have been removed entirely:
- sync test that took a log handler argument
- async test that took report_result & log_handler
- env that took report_result, env, log_handler params

We also do not recognize async and env test functions anymore - use
actors for those!