Fix .endswith() for short / long str
If the haystack str we are looking for the needle in is shorter than the
needle, we would calculate an incorrect starting position, potentially
pointing to uninitialized memory. Now explicitly check length to make
sure we stay within bounds of haystack str.

Added some test cases but since we often allocate memory
deterministically, I can't reproduce the issue I saw with this test.
However, the out-of-bounds problem that I saw when running `acton test`
no longer occurs in hundreds of test runs - it used to pop up in like
half the runs previously.