Test run MUSL execs on Linux
We now "cross-compile" for MUSL libc and test run the resulting binary.
We can only do this if we actually target the native CPU architecture,
so it's not cross-compilation per say, but testing MUSL on Linux. At one
point I think I had a MUSL binary that didn't run, so this allows us to
do a basic smoke test at least.