Avoid writing .ty in Types.reconstruct
Instead of letting Types.reconstruct directly write a .ty file, we
return the necessary information and write it from actonc instead. This
way, reconstruct is "pure" (not haskell pure) in the sense that it
doesn't have side effects of writing files to the file system - much
nicer for running when testing etc.