Recognize projects by Build.act & build.act.json
We now recognize Acton projects through the presence of either of:
- Acton.toml (existing)
- Build.act (new ideal!)
- build.act.json (commonly used now, but optional)

The future is the Build.act file but we're accepting multiple files in
the interrim.

Also check that there is a src/ directory in the same folder, in order
to robustify the project discovery a tiny bit. We are still susceptible
to finding a Build.act file "in the wrong place" like imagine we have
- foo/ - our Acton project
- foo/Build.act
- foo/src/
- foo/src/foo.act - just a source file
- foo/src/Build.act - it's allowed to have a Build.act IN your project
- foo/src/src/ - OK, this is a little weird, but possibly allowed

Running project discovery for foo.act would discover the foo/src/ as the
project root... I don't really know how to avoid this as long as we use
a single .act file to recognize an Acton project. Build.act (capital B)
makes it a little less likely to conflict since the convention is to use
lower case for module names, but it's not forbidden to call your own
module Build.act nor is it forbidden to have a subdirectory called src/

I opted for Build.act rather than build.act to reduce the likelihood of
conflicts but I also think capital B Build.act looks nice, similar to
Makefile, Dockerfile or Justfile. Makes it stand out a little bit.

Perhaps one strategy could be to search from the root and down instead
of from the file and up. Should we allow projects to be nested within
each other? Why is that valid? Feels weird... food for thought.