Allow logging of optional values
We can pass a data dict to logging functions, like:

    log.init("msg", data={"name": name})

The data argument is now a `dict[key, ?value]`, used to require that
every item passed in was a value. We now allow optional values, which is
just much more convenient as the user can feed in whatever they have,
not having to think about unwrapping the optional value. This work is
really the spiritual continuation to when we changed to str(?value).