Revamp parser exceptions
The FailFastErrors are really all from the parser, so now calling these
CustomParserException, which is a thin wrapper around CustomParserErrors
- this way we can treat errors the same regardless how they propagate to
us, via megaparsec errors or jumping direct using
Control.Exception (wrapped in CustomParserException). This means that we
now use nice custom error types for exceptions we raise as
Control.Exception from within the parser rather than just a single text
message as we had in the FailFastError. Much nicer and we can now
include hints in the parser exceptions!