Revamp class attribute init check
This is a rewrite of the class attribute initialization check. In the
__init__ method we have a constructor part which must unconditionally
initialize all attributes, so that we get a valid object. In the
constructor, references to `self` must not escape until after the
construction of the new object is complete (i.e. all attributes have
been initialized) nor can we call self.methods() since those methods
typically assume the self object is valid.

The inferrence of properties now scans through the entire __init__ method
looking for self.x assignments. Previously it would stop at some
particular statements, but that limited our view, so that's been removed
and we are now scanning the entirety of __init__.

We dynamically determine where the constructor ends by scanning through
the __init__ method up to the point where a reference to `self` is made
- at which point the constructor must be complete since we cannot pass a
reference to `self` before the own object has been properly constructed.
We make an exception for call self methods with NotImplemented bodies -
since those are implemented in C, we reckon the user also knows not to
touch the wrong self attributes. Loops are skipped past, since we cannot
statically determine if assignments in loops will happen or not,
however, we do scan to ensure that no references to self escape from
within the loop.

The analysis is more complete around handling of try / except as well as
if/else branches. Previously, we would only support:

  def __init__(self, foo):
      self.x = 1 if foo is not None else 3

Now we support:

  def __init__(self, foo):
      if foo is not None:
          self.x = 1
      else:
          self.x = 3

Similarly, we can mix in try / except

    def __init__(self, num : int, denom : int):
        if denom == 0:
            raise ValueError("Denominator in rational is zero")
        else:
            if denom > 0:
                self.num = num
                self.denom = denom
            else:
                self.num = -num
                self.denom = -denom

Added a page to acton.guide explaining most of this as well.