Revamp class attribute init check
This is a rewrite of the class attribute initialization check. In the
__init__ method we have a constructor part which must unconditionally
initialize all attributes, so that we get a valid object. In the
constructor, references to `self` must not escape until after the
construction of the new object is complete (i.e. all attributes have
been initialized) nor can we call self.methods() since those methods
typically assume the self object is valid.

The inferrence of properties now go through the entire __init__ method
looking for self.x assignments. It doesn't stop at any particular
statements, so this gives a better view of all attributes.

We dynamically determine where the constructor ends by scanning through
the __init__ method. Assignments are accepted, if / else are accepted as
part of the constructor as long as assignments are made in all branches.
Local variables are OK as are calling functions or methods on other
objects. `self` must not be referenced other than for the local
attributes, like self.x. Loops are not okay and will consider to be the
end of the constructor part since a loop might not run and thus won't
unconditionally initialize an assignment made within the loop.

The analysis is more complete around handling of try / except and
branches. Previously, we would only support:

  def __init__(self, foo):
      self.x = 1 if foo is not None else 3

Now we support:

  def __init__(self, foo):
      if foo is not None:
          self.x = 1
      else:
          self.x = 3

Similarly, we can mix in try / except

    def __init__(self, num : int, denom : int):
        if denom == 0:
            raise ValueError("Denominator in rational is zero")
        else:
            if denom > 0:
                self.num = num
                self.denom = denom
            else:
                self.num = -num
                self.denom = -denom

Added a page to acton.guide explaining most of this as well.